% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{alter_trib_inflows_for_reservoir}
\alias{alter_trib_inflows_for_reservoir}
\title{Alter tributary inflow file to reflect storage and releases from a reservoir}
\usage{
alter_trib_inflows_for_reservoir(
  sfr_component,
  output_dir,
  filename,
  verbose = TRUE
)
}
\arguments{
\item{sfr_component}{Dataframe of months and either: 1) subwatershed inflow partition
fractions (see \code{\link{gen_monthly_sfr_flow_partition}}), or 2) inflows available
for irrigation, or 3) inflows NOT available for irrigation (e.g. reserved for environmental
flows) (for items 2 and 3 see \code{\link{process_sfr_inflows}}).}

\item{output_dir}{directory to write the files to}

\item{filename}{Writes to this filename}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None, writes file to disk
}
\description{
Alter tributary inflow file to reflect storage and releases from a reservoir
}
\examples{
\dontrun{
subws_inflows <- process_sfr_inflows(scen, "daily_tributary_streamflow.txt")
write_SWBM_SFR_inflow_files(subws_inflows$non_irr, '.', "subwatershed_irrigation_inflows.txt")
}
}
