% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{assimilate_fj_update}
\alias{assimilate_fj_update}
\title{Assimilate Fort Jones Historical & New Updated Data}
\usage{
assimilate_fj_update(historical, fj_update)
}
\arguments{
\item{historical}{DataFrame of historical (1941 - ) data}

\item{fj_update}{DataFrame of more recent data}
}
\value{
DataFrame of single, combined dataset
}
\description{
Assumes that fj_update is the more accurate dataset (will drop historical days with overlap)
}
