% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWBMisc.R
\name{build_cdfw_instream_flow_cal}
\alias{build_cdfw_instream_flow_cal}
\title{Calculate SWBM Instream Flow Available Ratio Calendar}
\usage{
build_cdfw_instream_flow_cal(
  model_start_date,
  model_end_date,
  fort_jones_flows,
  cdfw_file = file.path(data_dir["ref_data_dir", "loc"], "cdfw_2017_instream_flows.csv")
)
}
\arguments{
\item{model_start_date}{Model start date}

\item{model_end_date}{Model end date}

\item{cdfw_file}{Existing CDFW file to calculate from, cdfw_2017_instream_flows.csv by default}
}
\value{
dataframe
}
\description{
Calculate SWBM Instream Flow Available Ratio Calendar
}
\author{
Claire Kouba
}
