% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression_old.R
\name{calc_mean_sd_before_date}
\alias{calc_mean_sd_before_date}
\title{Calculate Monthly Mean & Standard Deviation using data before a certain date}
\usage{
calc_mean_sd_before_date(df_daily_mean, stat_cutoff_date, Max_Num_Missing_Days)
}
\arguments{
\item{df_daily_mean}{dataframe of daily average flows (columns: Date, mean_cfs, Qual_Flag)}

\item{stat_cutoff_date}{date after which data should not be included in the mean and sd}
}
\value{

}
\description{
Very case-specific function created to ensure the means & std devs used for normalization do
not change over time (barring corrections to earlier data). Repeats many of the calculations
in RSVP::preproflow() but on a data subset.
}
