% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_num_stress_periods}
\alias{calc_num_stress_periods}
\title{Calculate Number of Stress Periods}
\usage{
calc_num_stress_periods(model_start, model_end, interval = "month")
}
\arguments{
\item{model_start}{Date of model start}

\item{model_end}{Date of model end}

\item{interval}{increment of the sequence, see seq.Date argument "by". "month" by default.}
}
\value{

}
\description{
Calculate Number of Stress Periods
}
