% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.R
\name{calc_split_sample_stats.grouped}
\alias{calc_split_sample_stats.grouped}
\title{Calculate Grouped Split Sampled Statistics}
\usage{
calc_split_sample_stats.grouped(
  dates,
  sim,
  obs,
  groups,
  split_date,
  FUNs,
  FUN_names
)
}
\arguments{
\item{dates}{Array of dates related to sim/obs}

\item{sim}{Array of simulated values}

\item{obs}{Array of observed values}

\item{groups}{Array of groups that sim, obs values belong to (same length as sim & obs)}

\item{split_date}{Date to split values on, based on dates array}

\item{FUNs}{Function(s) to calculate statistics using. Must accept arguments as (sim, obs)}

\item{FUN_names}{Names of functions in FUNs, used for DF column names}
}
\value{
Dataframe of statistics
}
\description{
Calculate Grouped Split Sampled Statistics
}
\examples{
\dontrun{
# Incomplete example
stream_combined <- ts_obs_sim_combine(streams, streams_sim, stream_names, val_cols = c('Flow','Flow_cfs'))
sfrcompare <- calc_split_sample_stats.grouped(stream_combined$Date, stream_combined$sim,
                                              stream_combined$obs, stream_combined$group,
                                              cal_split,
                                              FUNs = list(NSE, rmse, KGE),
                                              FUN_names = c('NSE', 'RMSE', 'KGE'))
}
}
\seealso{
\code{\link{calc_split_sample_stats}},
\code{\link{print.split_sample_stats}},
\code{\link{ts_obs_sim_combine}}
}
\author{
Leland Scantlebury
}
