% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelEval.R
\name{calc_split_sample_stats}
\alias{calc_split_sample_stats}
\title{Calculate Split Sample Statistics}
\usage{
calc_split_sample_stats(dates, sim, obs, split_date, FUNs, FUN_names)
}
\arguments{
\item{dates}{Array of dates related to sim/obs}

\item{sim}{Array of simulated values}

\item{obs}{Array of observed values}

\item{split_date}{Date to split values on, based on dates array}

\item{FUNs}{Function(s) to calculate statistics using. Must accept arguments as (sim, obs)}

\item{FUN_names}{Names of functions in FUNs, used for DF column names}
}
\value{
Dataframe of statistics
}
\description{
Calculate Split Sample Statistics
}
\examples{
\dontrun{
# Read in head observation data
#-- HOB data
hob_locs <- read.csv('hob_wells.csv', row.names=1, stringsAsFactors = F)
hob <- import_HOB(hob_input = 'SVIHM.hob',
                  hob_output = 'HobData_SVIHM.dat',
                  origin_date = origin_date)
hob <- hob[order(hob$row, hob$column),]

# Date to split data (i.e. pre-post calibration)
cal_split <- as.Date('2012-10-01')

hob_compare <- calc_split_sample_stats(dates = hob$date,
                                       sim = hob$sim,
                                       obs = hob$hobs,
                                       split_date = cal_split,
                                       FUNs = list(NSE, rmse, KGE),
                                       FUN_names = c('NSE', 'RMSE', 'KGE'))
}
}
\seealso{
\code{\link{calc_split_sample_stats.grouped}},
\code{\link{print.split_sample_stats}},
}
\author{
Leland Scantlebury
}
