% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{calc_station_dist_matrix}
\alias{calc_station_dist_matrix}
\title{Calculate station-to-station distance matrix}
\usage{
calc_station_dist_matrix(stations)
}
\arguments{
\item{stations}{DataFrame with station locations as columns latitude, longitude
and column name_short with short names for stations (row, col headers)}
}
\value{
Symmetric Matrix of distances between stations
}
\description{
Calculate station-to-station distance matrix
}
