% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{calc_swbm_spatial_precip_factors}
\alias{calc_swbm_spatial_precip_factors}
\title{Calculate SWBM Precipitation Factors}
\usage{
calc_swbm_spatial_precip_factors(
  poly_tab_filename = "polygons_table.txt",
  poly_shapefile_filename = NA,
  rainfall_records = NA,
  rainfall_locations = NA,
  verbose = TRUE
)
}
\arguments{
\item{rainfall_records}{Dataframe containing daily precipitation records (includes
public weather station data and date collected from private gauges).}

\item{rainfall_locations}{Dataframe containing locations for each rainfall record (includes
public weather station data and date collected from private gauges).}

\item{verbose}{T/F write status info to console (default: TRUE)}

\item{poly_filename}{Filename containing table of land use polygons}

\item{output_dir}{Directory to write file}
}
\value{
A two-column dataframe containing the SWBM polygon (field) ID; and the precip
multiplication factor.
}
\description{
Relates each landuse polygon (field) to spatially distributed rainfall data.
Generates a multiplication factor for each field to modify the model-domain-wide daily
precipitation input record.
}
\author{
Leland Scantlebury and Claire Kouba
}
