% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{combine_east_and_south_fork}
\alias{combine_east_and_south_fork}
\title{Combine East and South Fork streamflow records}
\usage{
combine_east_and_south_fork(tribs_list)
}
\arguments{
\item{tribs_list}{List of dataframes containing tributary flow data (gap-filled with FJ flow)}
}
\value{
List of daily mean dataframes (With E and S forks replaced by Scott River)
}
\description{
Reads in records from two main forks of Scott River and combines them for model input.
}
\details{
Input/Tributary Gauges included:
\itemize{
\item East Fork Scott River
\item South Fork Scott River
\item Sugar Creek
\item Etna Creek
\item French Creek
\item Patterson Creek
\item Kidder Creek
\item Moffett Creek
\item Mill Creek
\item Shackleford Creek
}

No Data for Clark, Johnson, Crystal, or Oro Fino Creeks.
}
