% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_scenario_dir}
\alias{create_scenario_dir}
\title{Create Scenario Directory}
\usage{
create_scenario_dir(dir_name)
}
\value{
new or existing update directory
}
\description{
For 02_ScenarioBuilder to create/check if scenario directory exists.
}
\examples{
create_update_dir()
}
