% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{create_SWBM_curtailment_df}
\alias{create_SWBM_curtailment_df}
\title{Create SWBM Curtailment Fractions File for SWBM}
\usage{
create_SWBM_curtailment_df(start_date, end_date, scenario_id)
}
\arguments{
\item{start_date}{Start date of the simulation (as a Date object).}

\item{end_date}{End date of the simulation (as a Date object).}

\item{scenario_id}{The curtailment scenario (anything but \code{basecase} returns a dataframe of zeros)}
}
\value{
A \code{data.frame} whose first column is \code{Stress_Period} (Date)
and whose remaining columns are named \code{ID_<fieldID>} with Curtailment fractions
for each field in each stress period.
}
\description{
Generates DataFrame water curtailment fractions for each SWBM agricultural
field for each stress period in the simulation. Fields without curtailments are assigned a default value of 0.0.
}
\details{
The curtailment fraction file is used in the SWBM model to specify reductions in applied irrigation water
due to local cooperative solutions (LCS) and regulatory curtailments. The fraction represents the proportion
of normal applied water that is reduced, with values typically ranging from 0.0 (no curtailment) to 1.0
(full curtailment).

When scenario_id is basecase, the function reads in the stored curtailment data for different
years (2021, 2022, 2023, and 2024), replaces overlapping dates, and ensures the final dataset is sorted
before saving to file.

Any other scenario_id results in a \code{data.frame} of 0.0 for every field/month.
}
\examples{
create_SWBM_curtailment_df(start_date = as.Date("2020-10-01"),
                           end_date = as.Date("2024-09-30"),
                           scenario_id = "basecase")
}
