% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_update_dir}
\alias{create_update_dir}
\title{Create Update Directory}
\usage{
create_update_dir(end_date)
}
\value{
new or existing update directory
}
\description{
For 01_InputDataUpdate.R to create/check if update directory exists. File formula is
SVIHM_Input_Files/Updates/\link{current date}
}
\examples{
create_update_dir()
}
