% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df.add_stats.byDate}
\alias{df.add_stats.byDate}
\title{Add basic stats (mean, sd), calculated over a time subset, to dataframe attributes}
\usage{
df.add_stats.byDate(
  df,
  value_col,
  date_col = "Date",
  start_date = NULL,
  end_date = NULL,
  na.rm = F
)
}
\arguments{
\item{df}{Data.frame of time series data}

\item{value_col}{Column name to calculate stats on}

\item{date_col}{Column with Date-type values}

\item{start_date}{Date at which to start the statistic calculation}

\item{end_date}{Date at which to end the statistic calculation}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}
}
\value{
Dataframe with added statistic attributes
}
\description{
Add basic stats (mean, sd), calculated over a time subset, to dataframe attributes
}
\author{
Leland Scantlebury
}
