% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df.add_stats}
\alias{df.add_stats}
\title{Add basic stats (mean, sd) to dataframe attributes}
\usage{
df.add_stats(df, value_col, na.rm = FALSE)
}
\arguments{
\item{df}{Data.frame of data}

\item{value_col}{Column name to calculate stats on}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds. (optional)}
}
\value{
Dataframe with added statistic attributes
}
\description{
Add basic stats (mean, sd) to dataframe attributes
}
\author{
Leland Scantlebury
}
