% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df.copy_stats}
\alias{df.copy_stats}
\title{Copy data frame basic stats}
\usage{
df.copy_stats(df1, df2, overwrite = T)
}
\arguments{
\item{df1}{Source of statistics dataframe}

\item{df2}{Dataframe to copy stats too}

\item{overwrite}{T/F, check is stats already exist before copying, don't copy if they exist
(default: True)}
}
\value{
dataframe with mean & sd as attributes
}
\description{
Copy data frame basic stats
}
\author{
Leland Scantlebury
}
