% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df.destandarize}
\alias{df.destandarize}
\title{De-Standarize Data}
\usage{
df.destandarize(df, value_col, new_col = NULL)
}
\arguments{
\item{df}{Dataframe, with attributes mean and sd related to value_col
(see \code{\link{df.add_stats}})}

\item{value_col}{Column name to denormalize}

\item{new_col}{Column to write destandardized data too (optional, by default will overwrite
value_col)}
}
\value{
dataframe with de-standarized data
}
\description{
De-standarizes data by multiplying by the standard deviation and adding the mean. Reverse of
df.standarize
}
\author{
Leland Scantlebury
}
