% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{df.standardarize}
\alias{df.standardarize}
\title{Standardize Data (z-score normalization)}
\usage{
df.standardarize(df, value_col)
}
\arguments{
\item{df}{Dataframe, with attributes mean and sd related to value_col
(see \code{\link{df.add_stats}})}

\item{value_col}{Column name to normalize}
}
\value{
dataframe with standardized value_col
}
\description{
Standardizes value_col by subtracting the mean and dividing by the standard deviation
}
\author{
Leland Scantlebury
}
