% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverData.R
\name{download_fort_jones_flow}
\alias{download_fort_jones_flow}
\title{Download Fort Jones Flow from USGS}
\usage{
download_fort_jones_flow(
  start_date,
  end_date,
  output_dir,
  save_csv = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{Start of data period}

\item{end_date}{End of data period}

\item{output_dir}{Directory to write file}

\item{save_csv}{boolean if csv of data should be saved to ref_data_dir, FALSE by default}

\item{verbose}{T/F write status info to console}
}
\value{
dataframe
}
\description{
Download Fort Jones Flow from USGS
}
