% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{download_meteo_data}
\alias{download_meteo_data}
\title{Download (Daily) Meteorological Data for Gauges from NOAA
Wrapper for rnoaa function meteo_pull_monitors. Converts data from tenths (mm, degrees) to full
value (i.e., from 1/10 mm to mm)}
\usage{
download_meteo_data(station_ids, start_date, end_date, datatypeid = "all")
}
\arguments{
\item{station_ids}{array of station ids to obtain data from}

\item{start_date}{Start of data period}

\item{end_date}{End of data period}

\item{datatypeid}{Accepts a valid data type id or a vector or list of data type ids. (optional)
(see rnoaa documentation, passed to rnoaa::meteo_pull_monitors as var)}
}
\value{
tibble of precip data (long data, for all monitors)
}
\description{
Download (Daily) Meteorological Data for Gauges from NOAA
Wrapper for rnoaa function meteo_pull_monitors. Converts data from tenths (mm, degrees) to full
value (i.e., from 1/10 mm to mm)
}
\examples{
noaa <- download_meteo_data(c("USC00041316", "USC00043182"),
                            start_date = '2012-10-01',
                            end_date = '2013-09-30',
                            datatypeid = "PRCP")
}
\author{
Leland Scantlebury
}
