% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverData.R
\name{download_USGS_flow}
\alias{download_USGS_flow}
\title{Download USGS Flow Data for Specified Gauge. Thin wrapper for \code{\link[dataRetrieval]{readNWISdv}}.}
\usage{
download_USGS_flow(site_no, start_date, end_date, parameterCd = "00060")
}
\arguments{
\item{site_no}{USGS Site Number}

\item{start_date}{Start of data period}

\item{end_date}{End of data period}

\item{parameterCd}{Parameter codes of data to download (daily flow=00060 by default)}
}
\value{
dataframe
}
\description{
Download USGS Flow Data for Specified Gauge. Thin wrapper for \code{\link[dataRetrieval]{readNWISdv}}.
}
\seealso{
\code{\link[dataRetrieval]{readNWISdv}}
}
