% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{fill_fj_cal_gv_gaps_regression_table}
\alias{fill_fj_cal_gv_gaps_regression_table}
\title{Fill Fort Jones, Cal, and Greenview Gaps using Regression}
\usage{
fill_fj_cal_gv_gaps_regression_table(
  model_coeff,
  station_table,
  daily_precip,
  start_date,
  end_date
)
}
\arguments{
\item{model_coeff}{Model coefficients generated using make_model_coeffs_table}

\item{station_table}{Dataframe of station info (i.e., data(noaa_stations))}

\item{daily_precip}{DataFrame with columns Date and PRCP_mm_XX, from make_daily_precip()}

\item{start_date}{Start of data period}

\item{end_date}{End of data period}
}
\description{
Use best available regression to fill in gaps in precip records for FJ and Cal
}
\author{
Claire Kouba
}
