% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_data_periods}
\alias{find_data_periods}
\title{Find periods of time series with data}
\usage{
find_data_periods(df, max_missing_days, date_col = "Date", value_col = "value")
}
\arguments{
\item{df}{Data.frame of time series data}

\item{max_missing_days}{Maximum number of days allowed to be missing before constituting a
separate data period}

\item{date_col}{Column with Date-type values}

\item{value_col}{Column with data whereupon non-NA values constitute data being present}
}
\value{
Data.frame with period starts and ends as dates
}
\description{
Find periods of time series with data
}
\examples{
# Load tributary data
daily_all <- read_gauge_daily_data()

# Subset to South Fork
sf <- df <- daily_all$South_Fork

# Get data periods where no more than a month of data is missing
sf_periods <- find_data_periods(sf, max_missing_days=30, value_col='mean_cfs')
}
