% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{gauge_prep}
\alias{gauge_prep}
\title{Title}
\usage{
gauge_prep(
  df,
  max_missing_days = 3,
  value_col = "AF",
  date_col = "Date",
  monthly = TRUE
)
}
\arguments{
\item{date_col}{}
}
\description{
Title
}
