% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{gauge_reg_obs_merge}
\alias{gauge_reg_obs_merge}
\title{Title}
\usage{
gauge_reg_obs_merge(regression_df, obs_df, obs_cols = c("AF"))
}
\arguments{
\item{obs_cols}{}
}
\value{

}
\description{
Title
}
