% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{gauge_regression_prep}
\alias{gauge_regression_prep}
\title{Title}
\usage{
gauge_regression_prep(
  gauge,
  stats_donor = NULL,
  start_date = NULL,
  end_date = NULL,
  value_col = "AF",
  date_col = "Date",
  transform = log10,
  col_rename = "normLogAF"
)
}
\description{
Title
}
