% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWBM_crop_coefficents.R
\name{gen_daily_binary_crop_coefficients}
\alias{gen_daily_binary_crop_coefficients}
\title{Generate Daily Binary Crop Coefficients (Kc)}
\usage{
gen_daily_binary_crop_coefficients(
  model_start_date,
  model_end_date,
  kc_dormant = 0,
  kc_growing = 0.9,
  growing_season_start_month = 3,
  growing_season_start_day = 1,
  growing_season_end_month = 11,
  growing_season_end_day = 14,
  irr_demand_mult = 1
)
}
\arguments{
\item{model_start_date}{Start date of model}

\item{model_end_date}{End date of model}

\item{kc_dormant}{coefficient during dormant period, default 0.0}

\item{kc_growing}{coefficient during growing period, default 0.9}

\item{growing_season_start_month}{growing season start month, default 3 (March)}

\item{growing_season_start_day}{growing season deafult day, default 1 (March 1st)}

\item{growing_season_end_month}{growing season start month, default 11 (November)}

\item{growing_season_end_day}{growing season deafult day, default 14 (November 14th)}

\item{irr_demand_mult}{Irrigation demand multiplier, default 1}
}
\description{
Currently used for alfalfa and pasture. Binary coefficients for growing/dormant.
}
\author{
Claire Kouba, Leland Scantlebury
}
