% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWBM_crop_coefficents.R
\name{gen_daily_curve_crop_coefficients}
\alias{gen_daily_curve_crop_coefficients}
\title{Generate Daily Curve Crop Coefficients (Kc)}
\usage{
gen_daily_curve_crop_coefficients(
  model_start_date,
  model_end_date,
  kc_dormant = 0,
  kc_by_crop_stage = c(0, 0.27, 1.15, 1.15, 0),
  days_in_crop_stage = c(38, 26, 32, 36),
  growing_season_start_month = 3,
  growing_season_start_day = 14,
  ref_data_dir = data_dir["ref_data_dir", "loc"],
  legacy_file = "kc_grain_1991_2011.txt",
  irr_demand_mult = 1
)
}
\arguments{
\item{model_start_date}{Start date of model}

\item{model_end_date}{End date of model}

\item{kc_dormant}{coefficient during dormant period, default 0.0}

\item{kc_by_crop_stage}{array of coefficients for various growing stages (default: grain curve)}

\item{days_in_crop_stage}{array of days for each stage (2 less than kc_by_crop_stage in length)}

\item{growing_season_start_month}{growing season start month, default 3 (March)}

\item{growing_season_start_day}{growing season deafult day, default 1 (March 14th)}

\item{ref_data_dir}{Reference data directory (default: SVIHM/SVIHM_Input_Files/reference_data/)
used to fine legacy_file}

\item{legacy_file}{legacy crop coefficient file to overwrite, default kc_1991_2011.txt. NULL
to ignore}

\item{irr_demand_mult}{Irrigation demand multiplier, default 1}
}
\value{

}
\description{
Currently used for grain (see: default values). Uses a curve to represent crop stages.
}
\author{
Claire Kouba, Leland Scantlebury
}
