% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWBM_SFR_files.R
\name{gen_sfr_flow_partition}
\alias{gen_sfr_flow_partition}
\title{Generate Monthly SFR Inflow Partition}
\usage{
gen_sfr_flow_partition(
  model_start_date,
  model_end_date,
  update_dir,
  monthly = T,
  streamflow_records_file = "streamflow_records_regressed.txt",
  unchanging_partition = F
)
}
\arguments{
\item{model_start_date}{Start date of model}

\item{model_end_date}{End date of model}

\item{unchanging_partition}{Are the inflow partitioning values unchanging over time? True or false}
}
\description{
Inflow to the model domain is often measured downstream of the domain boundary.
The upstream inflow may enter the domain on multiple forks/inflow points.
This partitioning assigns a fraction of the flow in each subwatershed to
different inflow points. Inflow points are assigned to subwatersheds in file
SFR_inflow_segments.txt.
}
\author{
Claire Kouba, Leland Scantlebury
}
