% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression_old.R
\name{generate_streamflow_input}
\alias{generate_streamflow_input}
\title{Read in River Gauge Daily Means}
\usage{
generate_streamflow_input(
  end_date,
  fj_update = NULL,
  sf_reg_dir = data_dir["sf_reg_dir", "loc"],
  max_missing_days = 3,
  stat_cutoff_date
)
}
\arguments{
\item{end_date}{End of data period}

\item{fj_update}{DataFrame of latest Fort Jones (FJ data), from download_fort_jones_flow() (optional)}

\item{sf_reg_dir}{streamflow regression directory (optional, default: SVIHM/Streamflow_Regression_Model)}

\item{max_missing_days}{Maximum number of missing daily flow data before month is removed from
the regression (optional, default = 3)}

\item{historical}{DataFrame of historical (1941 - ) data}
}
\value{
List of daily mean dataframes (FJ followed by order above)

DataFrame of single, combined dataset
}
\description{
Reads in primary gauge (Fort Jones) and input/tributary gauge daily.
}
\details{
Input/Tributary Gauges included:
\itemize{
\item East Fork Scott River
\item South Fork Scott River
\item Sugar Creek
\item Etna Creek
\item French Creek
\item Patterson Creek
\item Kidder Creek
\item Moffett Creek
\item Mill Creek
\item Shackleford Creek
}

No Data for Clark, Johnson, Crystal, or Oro Fino Creeks.
}
\examples{
Assimilate Fort Jones Historical & New Updated Data

Assumes that fj_update is the more accurate dataset (will drop historical days with overlap)

Generate SVIHM Streamflow Inputs

}
