% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{get_daily_precip_table}
\alias{get_daily_precip_table}
\title{Get table of daily precipitation}
\usage{
get_daily_precip_table(
  start_date,
  end_date,
  ref_data_dir = data_dir["ref_data_dir", "loc"],
  na_fill = "average",
  use_corrected_fj = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{start_date}{Start of data period}

\item{end_date}{End of data period}

\item{ref_data_dir}{Reference data directory
(optional, default: SVIHM/SVIHM_Input_Files/reference_data)}

\item{na_fill}{string of NA filling behavior, 'average' for averages of the preceding and next-day-with-data values,
pass a numeric value to fill NA with that value, or NA for no filling (default: 'average')}

\item{use_corrected_fj}{T/F use local file to correct Fort Jones (FJ) precip data (NOAA API has missing values present in file)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\description{
Get table of daily precipitation
}
\author{
Claire Kouba, Leland Scantlebury
}
