% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{get_fj_flow_file_name}
\alias{get_fj_flow_file_name}
\title{Get Fort Jones Flow File Name}
\usage{
get_fj_flow_file_name(start_date, end_date)
}
\arguments{
\item{start_date}{Start date of data}

\item{end_date}{End date of data}
}
\description{
Codifies a pattern for naming
}
\examples{
fname <- get_fj_flow_file_name(start_date = as.Date('2015-10-01'),
                               end_date = as.Date('2016-9-30'))

}
\author{
Leland Scantlebury
}
