% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_model_end}
\alias{get_model_end}
\title{Get Model End Date}
\usage{
get_model_end(end_year, water_years = TRUE)
}
\arguments{
\item{end_year}{Integer year}

\item{water_years}{Boolean}
}
\value{

}
\description{
Get Model End Date
}
