% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_model_start}
\alias{get_model_start}
\title{Get Model Start Date}
\usage{
get_model_start(start_year, water_years = TRUE)
}
\arguments{
\item{start_year}{Integer year}

\item{water_years}{Boolean}
}
\description{
Get Model Start Date
}
