% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMODFLOW.R
\name{import_HOB}
\alias{import_HOB}
\title{Import MODFLOW Head Observation (HOB) Files}
\usage{
import_HOB(hob_input, hob_output, origin_date)
}
\arguments{
\item{hob_input}{file path to MODFLOW HOB file}

\item{hob_output}{file path to MODFLOW HOB output}

\item{origin_date}{date from which to calculate calendar date from HOB file SP/TS offsets
(Currently assumes stress periods are months, time steps days)}
}
\value{
dataframe of obs/sim GW levels
}
\description{
Import MODFLOW Head Observation (HOB) Files
}
\author{
Leland Scantlebury
}
