% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMODFLOW.R
\name{import_sfr_gauge}
\alias{import_sfr_gauge}
\title{Import Streamflow Routing Package Gage Output File}
\usage{
import_sfr_gauge(gag_out, origin_date)
}
\arguments{
\item{gag_out}{file path to gage file}

\item{origin_date}{date from which to calculate calendar date from SFR time}
}
\value{
dataframe of gage data
}
\description{
Import Streamflow Routing Package Gage Output File
}
\author{
Leland Scantlebury
}
