% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{latest_dir}
\alias{latest_dir}
\title{Return Latest (most recent) Directory in a Directory}
\usage{
latest_dir(dir)
}
\arguments{
\item{dir}{path to directory to search through}
}
\value{
Latest directory full filepath
}
\description{
Return Latest (most recent) Directory in a Directory
}
\author{
Leland Scantlebury
}
