% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{make_model_coeffs_table}
\alias{make_model_coeffs_table}
\title{Create Precipitation Gauge Linear Regression Coefficient Table}
\usage{
make_model_coeffs_table(
  months = 1:12,
  station_table,
  daily_precip,
  ys = c("fj", "cal"),
  xs = c("fj", "cal", "gv", "et", "yr", "y2")
)
}
\arguments{
\item{months}{numeric months to be used in regression}

\item{daily_precip}{DataFrame with columns Date and PRCP_mm_XX, from make_daily_precip()}

\item{ys}{list of station abbreviations for predicting at}

\item{xs}{list of station abbreviations used to predict at ys}
}
\value{
DataFrame of coeffients
}
\description{
Performs regressions on the various gauges (columns) of the daily_precip table to obtain linear
coefficents for gauge combinations between those in xs and those in ys.
}
\details{
Expects the xs and ys to be datasets in daily_precip under the columns of 'PRCP_mm_(xs/ys)'
}
\author{
Claire Kouba, Leland Scantlebury
}
