% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mftime2date}
\alias{mftime2date}
\title{MODFLOW Model Time to Date}
\usage{
mftime2date(sp, ts, origin_date)
}
\arguments{
\item{sp}{Stress period}

\item{ts}{Time Step}

\item{origin_date}{Date at ts=0}
}
\description{
Assumes sp in months, and ts in days! Rounds to nearest day to prevent minor numerical errors
}
\author{
Leland Scantlebury
}
