% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWBM_SFR_files.R
\name{move_inflows}
\alias{move_inflows}
\title{Move inflows from one dataframe to another}
\usage{
move_inflows(inflow_to, inflow_from, date_start, date_end, streams = "ALL")
}
\arguments{
\item{date_start}{inclusive}

\item{date_end}{inclusive}

\item{streams}{}
}
\description{
Move inflows from one dataframe to another
}
