% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_stations.R
\docType{data}
\name{noaa_stations}
\alias{noaa_stations}
\title{Basic data on NOAA gauges used in SVIHM}
\format{
A data frame with 6 observations of 7 variables
}
\source{
rnoaa package
}
\usage{
data(noaa_stations)
\describe{
  \item{id}{NOAA station id}
  \item{name}{Common name of station}
  \item{name_short}{short 2-3 letter version of name}
  \item{noaa_name}{NOAA official name of station}
  \item{latitude}{latitude of station}
  \item{longitude}{longitude of station}
  \item{elevation}{elevation of station}
}
}
\description{
Location and name data for the noaa weather gauges used for the Scott Valley Integrated Hydrologic Model (SVIHM)
}
\keyword{datasets}
