% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.gw.hydrograph_wMap}
\alias{plot.gw.hydrograph_wMap}
\title{Plot Hydrograph w/ Map of Scott Valley}
\usage{
\method{plot}{gw.hydrograph_wMap}(
  dates,
  obs,
  sim,
  xloc,
  yloc,
  map_xs,
  map_ys,
  ylabel,
  log = "",
  title = NULL,
  colors = NULL,
  map_x_offset = 0,
  map_y_offset = 0,
  mapbmar = 30,
  ...
)
}
\arguments{
\item{dates}{Array of dates, or two arrays of dates corresponding to obs & sim}

\item{obs}{Array of observed data, plotted as points}

\item{sim}{Array of simulated (modeled) data, plotted as a line}

\item{xloc}{real-world map x coordinate of location being plotted (UTM Zone 10)}

\item{yloc}{real-world map y coordinate of location being plotted (UTM Zone 10)}

\item{map_xs}{real-world map x coordinates of all data locations (UTM Zone 10)}

\item{map_ys}{real-world map y coordinates of all data locations (UTM Zone 10)}

\item{ylabel}{Y-axis label}

\item{log}{'y' to log10 transform y-axis (default linear = '')}

\item{title}{Plot title}

\item{colors}{Colors for obs & sim data, respectively}

\item{map_x_offset}{x Buffer around map points (optional, default 0)}

\item{map_y_offset}{y Buffer around map points (optional, default 0)}

\item{mapbmar}{Margins on map bottom, used to move it up and down (and possibly avoid margin
errors) (optional, default 30)}

\item{...}{extra options passed to \code{\link{plot.ts_setup}}}
}
\value{
Plot
}
\description{
Plot Hydrograph w/ Map of Scott Valley
}
\examples{
# Real locations
hob_locs <- read.csv(file.path(data_dir['ref_data_dir','loc'], 'hob_wells.csv'),
                     row.names=1, stringsAsFactors = F)
locx <- hob_locs[1,'x']
locy <- hob_locs[1,'y']

# Fake Data
dates <- seq.Date(as.Date('1991-10-01'), as.Date('2001-9-30'), 'month')
obs <- runif(length(dates), 5, 50)
sim <- obs - runif(length(dates), -5, 5)

# May have to print to PDF to avoid figure margins error
\dontrun{
  pdf('plot_gw_hydrograph_wMap_example.pdf', width=11, height=8.5)
  plot.gw.hydrograph_wMap(dates, obs, sim,
                          locx, locy, hob_locs[,'x'], hob_locs[,'y'],
                          ylabel='Y', title='Water Levels at Loc 1')
  dev.off()
}
}
\author{
Leland Scantlebury
}
