% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pre_post_compare}
\alias{plot.pre_post_compare}
\title{Compare Time Series Data Before and After A Specified Date}
\usage{
\method{plot}{pre_post_compare}(
  dates,
  obs,
  sim,
  split_date,
  ylabel,
  log = "",
  pre_title = NULL,
  post_title = NULL,
  colors = NULL,
  zero_rep = 0.1,
  ...
)
}
\arguments{
\item{dates}{Array of dates (must be same for obs & sim)}

\item{obs}{Array of observed data}

\item{sim}{Array of simulated (modeled) data}

\item{split_date}{Date to split values on, based on dates array}

\item{ylabel}{Y-axis label}

\item{log}{'y' to log10 transform y-axis (default linear = '')}

\item{pre_title}{Title for pre split_date period}

\item{post_title}{Title for post split_date period}

\item{colors}{Colors for obs and sim data}

\item{zero_rep}{value to replace zeroes with (optional, only used with log='y')}

\item{...}{extra options passed to \code{\link{plot.ts_setup}}}
}
\value{
Plot
}
\description{
Intended for comparing pre-post calibration periods
}
\examples{
dates <- seq.Date(as.Date('1991-10-01'), as.Date('2001-9-30'), 'month')
obs <- runif(length(dates), 1, 50)
sim <- obs - runif(length(dates), -7, 5)
sim[sim < 0] <- 0

plot.pre_post_compare(dates, obs, sim, split_date = as.Date('1995-9-30'),
                      ylabel='Streamflow [m^3]',
                      pre_title = 'Pre-WY1995', post_title = 'Post-WY1995')
}
\author{
Leland Scantlebury
}
