% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.precip.compare}
\alias{plot.precip.compare}
\title{Precipitation Comparison Bar Chart}
\usage{
\method{plot}{precip.compare}(
  dates,
  precip1,
  precip2,
  unit = "mm",
  title = NULL,
  colors = NULL,
  legend_names
)
}
\arguments{
\item{dates}{array of dates of precip events}

\item{precip1}{array of first precipitation dataset at dates}

\item{precip2}{array of second precipitation dataset at dates}

\item{unit}{character, the units the precipitation is measured in}

\item{title}{title of plot (optional)}

\item{colors}{colors for data (identical, precip1, precip2) (optional)}

\item{legend_names}{names added to legend to describe precip1 and precip2}
}
\value{
Plot
}
\description{
Creates plot showing identical (color 1) and differing (color 2, color 3) precipitation data
}
\examples{
\dontrun{
# Read
precip1 <- read.table('../../SVIHM_Input_Files/Scenario_Development/precip_regressed_orig.txt',
                      header = F, col.names = c('prcp','date'))
precip2 <- read.table('../../SVIHM_Input_Files/Scenario_Development/precip_regressed.txt',
                      header = F, col.names = c('prcp','date'))
precip1$date <- as.Date(precip1$date, format='\%d/\%m/\%Y')
precip2$date <- as.Date(precip2$date, format='\%d/\%m/\%Y')

# Subset
precip1 <- precip1[precip1$date > '2011-9-30',]
precip2 <- precip2[precip2$date > '2011-9-30',]
precip2 <- precip2[precip2$date <= max(precip1$date),]

# Plot
plot.precip.compare(precip1$date,
                    precip1$prcp,
                    precip2$prcp,
                    unit = 'm',
                    legend_names = c('Original','New'))
}
}
\author{
Leland Scantlebury
}
