% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.precip}
\alias{plot.precip}
\title{Precipitation Bar Chart}
\usage{
\method{plot}{precip}(dates, precip, unit = "mm", title = NULL, col = NULL, border = par("fg"))
}
\arguments{
\item{dates}{array of dates of precip events}

\item{precip}{array of precipitation amount at dates}

\item{unit}{character, the units the precipitation is measured in}

\item{title}{title of plot (optional)}

\item{col}{color of bars (optional)}

\item{border}{color of bar borders (optional)}
}
\value{

}
\description{
Precipitation Bar Chart
}
