% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.scatterbox}
\alias{plot.scatterbox}
\title{1:1 Scatterplot w/ Boxplots on Top and Bottom}
\usage{
\method{plot}{scatterbox}(
  obs,
  sim,
  groups = NA,
  xlab = "Observed Values",
  ylab = "Simulated Values",
  title = NA,
  pch = 1,
  leg_cex = 1,
  col = NULL,
  log = F,
  zero_rep = 0.1
)
}
\arguments{
\item{obs}{Array of observed data}

\item{sim}{Array of simulated (modeled) data}

\item{groups}{Array of groups that sim, obs values belong to (optional, or same length as sim & obs)}

\item{xlab}{X-axis label (default 'Observed Values')}

\item{ylab}{Y-axis label (default 'Simulated Values')}

\item{pch}{Plotting 'character', i.e., symbol to use. See \code{\link[base]{points}} (default 1)}

\item{leg_cex}{Legend size parameter cex (default 1.0)}

\item{col}{Colors for groups}

\item{log}{T/F whether to log transform axes}

\item{zero_rep}{value to replace zeroes with (optional, only used with log=T)}
}
\value{
Plot
}
\description{
Used to show the fit of simulated data to observed data, with boxplots to account for the
density of the data
}
\examples{
# Fake Data
obs <- runif(250, 1, 10)
sim <- obs - runif(250, -2, 2)
sim[sim < 0] <- 0

# No Groups
plot.scatterbox(obs, sim)

# With Groups
groups <- rep('Group1',250)
groups[obs*sim > 55] <- 'Group2'
plot.scatterbox(obs, sim, groups = groups)
}
\author{
Leland Scantlebury
}
