% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.svihm_minimap}
\alias{plot.svihm_minimap}
\title{SVIHM Minimap Plot}
\usage{
\method{plot}{svihm_minimap}(xloc, yloc, map_xs, map_ys, map_x_offset, map_y_offset)
}
\arguments{
\item{xloc}{numeric UTM 10N x-coordinate of highlighted location}

\item{yloc}{numeric UTM 10N x-coordinate of highlighted location}

\item{map_xs}{numeric UTM 10N x-coordinates of locations}

\item{map_ys}{numeric UTM 10N x-coordinates of locations}

\item{map_x_offset}{x Buffer around map points}

\item{map_y_offset}{y Buffer around map points}
}
\value{
Plot
}
\description{
Plots coordinate data, with a specific highlighted coordinate, on a map of the SVIHM model with
the SFR river shown. Intended to be a minimap in the corner of a time series (etc) plot
}
\examples{
hob_locs <- read.csv(file.path(data_dir['ref_data_dir','loc'], 'hob_wells.csv'),
                     row.names=1, stringsAsFactors = F)
locx <- hob_locs[5,'x']
locy <- hob_locs[5,'y']

plot.svihm_minimap(locx, locy, hob_locs[,'x'], hob_locs[,'y'], map_x_offset = 0, map_y_offset = 0)
}
\seealso{
\code{\link{plot.gw.hydrograph_wMap}},\code{\link{plot.stream.hydrograph_wMap}}
}
\author{
Leland Scantlebury
}
