% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ts_setup}
\alias{plot.ts_setup}
\title{Setup Time Series Plot}
\usage{
\method{plot}{ts_setup}(
  dates,
  ...,
  xlabel,
  ylabel,
  log = "",
  interval = "year",
  bgcolor = "grey90",
  gridcolor = "white",
  gridcolor2 = "grey93",
  las = 2,
  xlim_override = NULL,
  ylim_min_diff = 10
)
}
\arguments{
\item{dates}{Array of dates, or list of arrays (for multiple datasets passed to ...)
(only used to determine axis limits)}

\item{...}{Array(s) of data to be plotted (only used to determine axis limits)}

\item{xlabel}{X-axis label}

\item{ylabel}{Y-axis label}

\item{log}{'y' to log10 transform y-axis (default linear = '')}

\item{interval}{time interval to draw x-axis grid at (optional, default 'year')}

\item{bgcolor}{background color of plot (optional, default 'grey90')}

\item{gridcolor}{color of major gridlines (optional, default 'white')}

\item{gridcolor2}{color of minor gridlines (only used for log plots) (optional, default 'grey93')}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. See \code{\link[base]{par}} (default 2)}

\item{xlim_override}{overrides x axis limits with provided values (optional)}

\item{ylim_min_diff}{overrides y axis limits with provided values (optional)}
}
\value{
Plot Setup
}
\description{
Sets up background, grid, and axes for a times series plot. Plots no actual data
}
\examples{
# Fake data
dates <- seq.Date(as.Date('1991-10-01'), as.Date('2001-9-30'), 'month')
ys <- runif(length(dates), 5, 50)

# Plot
plot.ts_setup(dates, ys, xlabel='X', ylabel='Y')
}
\seealso{
\code{\link{plot.gw.hydrograph_wMap}},\code{\link{plot.stream.hydrograph_wMap}}
}
\author{
Leland Scantlebury
}
