% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWBM_SFR_files.R
\name{process_sfr_inflows}
\alias{process_sfr_inflows}
\title{Process SFR Irrigation Inflows}
\usage{
process_sfr_inflows(
  model_start_date,
  model_end_date,
  scenario_id,
  stream_inflow_filename,
  avail_for_irr = T,
  instream_flow_regime = NA
)
}
\arguments{
\item{model_start_date}{Start date of model}

\item{model_end_date}{End date of model}

\item{stream_inflow_filename}{Filename of the gap-filled tributary flow records for each
subwatershed (see \code{\link{write_tributary_input_file}}).}

\item{avail_for_irr}{Boolean. Writes the file "subwatershed_irrigation_inflows.txt" if TRUE
and the file "subwatershed_nonirrigation_inflows.txt" if FALSE.}

\item{instream_flow_regime}{Date frame of start- and end-dates and flow values
describing an annual instream flow regime. If NA, produces a 0-flow nonirrigation inflows
table or passes 100\% of inflow to the irrigation inflows table.}

\item{instream_flow_units}{Handles "cfs", "cms" or "m3day". Defaults to NA.}
}
\description{
Generates table of monthly stream inflow volumes that are either available for irrigation
purposes or are unavailable for irrigation (i.e., flows reserved for environmental uses).
}
\author{
Claire Kouba, Leland Scantlebury
}
