% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{read_gauge_daily_data}
\alias{read_gauge_daily_data}
\title{Read in River Gauge Daily Means}
\usage{
read_gauge_daily_data(sf_reg_dir = data_dir["sf_reg_dir", "loc"])
}
\arguments{
\item{sf_reg_dir}{streamflow regression directory (optional, default: SVIHM/Streamflow_Regression_Model)}
}
\value{
List of daily mean dataframes (FJ followed by order above)
}
\description{
Reads in primary gauge (Fort Jones) and input/tributary gauge daily. Adds Acre-ft (AF) column
(calculated from cfs)
}
\details{
Input/Tributary Gauges included:
\itemize{
\item East Fork Scott River
\item South Fork Scott River
\item Sugar Creek
\item Etna Creek
\item French Creek
\item Patterson Creek
\item Kidder Creek
\item Moffett Creek
\item Mill Creek
\item Shackleford Creek
}

No Data for Clark, Johnson, Crystal, or Oro Fino Creeks.
}
