% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{resample2monthly}
\alias{resample2monthly}
\title{Resample Column to Monthly}
\usage{
resample2monthly(df, value_col, date_col = "Date", FUN, ...)
}
\arguments{
\item{df}{Dataframe to resample}

\item{value_col}{Column containing data to resample}

\item{date_col}{Date column of df (optional, assumes "Date")}

\item{FUN}{a function to compute the summary statistics}

\item{...}{further arguments passed to aggregate}
}
\value{
Dataframe resampled by FUN to monthly interval
}
\description{
Resamples a column of a dataset to monthly, given function FUN. Returns montly dataset with
date column (of same original name) with the first date of the aggregated month.
}
\author{
Leland Scantlebury
}
