% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_dev.R
\name{save_scen_param_file}
\alias{save_scen_param_file}
\title{Save Selected Scenario Parameters in a CSV}
\usage{
save_scen_param_file(scen, working_dir)
}
\arguments{
\item{scen}{A named list representing the scenario to be configured. Must include:
\itemize{
\item \code{name} - character; scenario name (used for output directory naming).
\item \code{type} - character; scenario type: one of \code{"BASECASE"}, \code{"UPDATE"}, or \code{"PRMS"} (case-insensitive).
}}
}
\value{
None. Saves a .csv file in the working directory.
}
\description{
Stores parameters in a consistent CSV format to facilitate tabular inter-
scenario comparisons. Writes the CSV file to the location specified
in \code{working_directory}.
}
\examples{
save_scen_param_file(scen, working_dir)
}
