% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWBM_SFR_files.R
\name{set_inflows}
\alias{set_inflows}
\title{Set inflows between dates to specified value}
\usage{
set_inflows(inflow_df, date_start, date_end, value, streams = "ALL")
}
\arguments{
\item{date_start}{inclusive}

\item{date_end}{inclusive}

\item{streams}{}
}
\description{
Set inflows between dates to specified value
}
