% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{streamflow_maps}
\alias{streamflow_maps}
\title{Generate Streamflow Maps}
\usage{
streamflow_maps(
  mf_dir,
  out_dir,
  stress_periods = NULL,
  save_as_pdf = TRUE,
  pdf_name = "wet_dry_stream_flipbook.pdf",
  flow_breaks_manual = c(0, 2.5, 20, 50, 100, 300, 700, 6350) * 1000,
  n_classes = 7
)
}
\arguments{
\item{mf_dir}{Character string. Directory path where the MODFLOW files are located.}

\item{out_dir}{Character string. Directory path where the output maps should be saved.}

\item{stress_periods}{Numeric vector. Optional. A vector of stress period indices to plot. If NULL, all stress periods will be plotted. Defaults to NULL.}

\item{save_as_pdf}{Logical. If TRUE, saves the output maps as a single PDF file. If FALSE, saves each stress period as a separate PNG file. Defaults to TRUE.}

\item{pdf_name}{Character string. The name of the output PDF file. Only used if \code{save_as_pdf} is TRUE. Defaults to "wet_dry_stream_flipbook.pdf".}

\item{flow_breaks_manual}{Numeric vector. Manual breaks for flow rate classification (in m3/day). Defaults to c(0, 2.5, 20, 50, 100, 300, 700, 6350) * 1000.}

\item{n_classes}{Integer. Number of flow classes to use for color breaks in the map. Defaults to 7.}
}
\value{
No return value. The function generates and saves the map files.
}
\description{
This function generates streamflow maps for selected stress periods from MODFLOW output. The streamflow data is processed, and a series of maps are generated showing the flow and depth of water in stream reaches, with flexible output options for PDF or PNG files.
}
\details{
This function reads the streamflow data from the MODFLOW global output file ("Streamflow_Global.dat") and processes the data to create a series of maps. It also reads the corresponding GIS data for stream segments and groundwater basin boundaries. The streamflow and depth data are plotted on maps with color coding based on user-defined breaks for flow rates. The output can either be a single PDF file or individual PNG images for each stress period.

The function also allows users to select which stress periods to plot, with the option to plot all stress periods if none are specified. The flow breaks for color classification can be manually specified, and the number of flow classes for the color palette can also be adjusted.

The function uses the \code{data_dir} dataframe internally to retrieve the path to the plot data directory.
}
\examples{
\dontrun{
# Example usage:
streamflow_maps(mf_dir = 'path_to_modflow_dir', out_dir = 'path_to_output_dir')
}

}
