% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subset.DateTwoSided}
\alias{subset.DateTwoSided}
\title{Two-Sided date subset}
\usage{
\method{subset}{DateTwoSided}(df, start = NULL, end = NULL, date_col = "Date", include_end = F)
}
\arguments{
\item{df}{Data.frame of time series data}

\item{start}{Date to start subset (inclusize)}

\item{end}{Date to end subset (exclusive, unless include_end=T)}

\item{date_col}{Column with Date-type values}

\item{include_end}{T/F switch if end date is inclusive (less than or equal to, default=F)}
}
\value{

}
\description{
Two-Sided date subset
}
