% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{swbm_build_field_value_df}
\alias{swbm_build_field_value_df}
\title{Build Field-value Data Frame}
\usage{
swbm_build_field_value_df(
  nfields = 2119,
  model_start_date,
  model_end_date,
  default_values = NA
)
}
\arguments{
\item{nfields}{number of fields in the Soil Water Balance Model simulation}

\item{model_start_date}{Start date of simulation}

\item{model_end_date}{End date of simulation}

\item{default_values}{Default value(s), given by field (optional, defaults to NA)}
}
\value{
Dataframe of stress periods (rows) for each field (columns)
}
\description{
Empty dataframe of field-level values, for every month between the start and end dates. This
format is common to many temporal SWBM inputs
}
\examples{
# Dates
start_date <- get_model_start(1991)
end_date <- as.Date(floor_date(Sys.Date(), 'month')-1)
# Fields
nfields <- 100
# For example: land use
default_lu <- rep(swbm_lutype['Alfalfa','Code'], nfields)

lu_df <- swbm_build_field_value_df(nfields, start_date, end_date, default_lu)

}
