% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{swbm_irrtype_cp_update}
\alias{swbm_irrtype_cp_update}
\title{Update Irrigation Type Table with New Data}
\usage{
swbm_irrtype_cp_update(irrtype_df, update_table, verbose = TRUE)
}
\arguments{
\item{irrtype_df}{Dataframe of field irrigation types for each stress period, possibly created by
\code{\link[=swbm_build_field_value_df]{swbm_build_field_value_df()}}}

\item{update_table}{Dataframe with columns 'ID' of field IDs and 'Year' of the year (integer XXXX)
when the field switched to center pivot. Zeros can be used to denote that it
has never switched.}

\item{verbose}{T/F write # of yearly changed to console (default: TRUE)}
}
\value{
Updated irrtype_df
}
\description{
Update Irrigation Type Table with New Data
}
\examples{
# Dates
start_date <- get_model_start(1991)
end_date <- as.Date(floor_date(Sys.Date(), 'month')-1)
# Fields
nfields <- 100
# For example: land use
default_irr <- rep(swbm_irrtype['Wheel Line','Code'], nfields)

lu_df <- swbm_build_field_value_df(nfields, start_date, end_date, default_irr)

# Make update table - magically all fields updated to center pivot in 2005!
updates <- data.frame('ID'=1:nfields, 'Year'=2005)

# Do updates
lu_df <- swbm_irrtype_cp_update(lu_df, updates)

}
