% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{swbm_landuse_update}
\alias{swbm_landuse_update}
\title{Update Land Use Table With New Data}
\usage{
swbm_landuse_update(lu_df, update_table, verbose = TRUE)
}
\arguments{
\item{lu_df}{Dataframe of field land uses for each stress period, possibly created by
\code{\link[=swbm_build_field_value_df]{swbm_build_field_value_df()}}}

\item{update_table}{Table of land use updates, where each column has a name of the form
'landuse_XX' where XX is year 20XX, each row is a field, in order, and the
cells contain valid land use codes (see \link{swbm_lutype})}

\item{verbose}{T/F write # of yearly changed to console (default: TRUE)}
}
\value{
Updated lu_df
}
\description{
Update Land Use Table With New Data
}
\examples{
# Dates
start_date <- get_model_start(1991)
end_date <- as.Date(floor_date(Sys.Date(), 'month')-1)
# Fields
nfields <- 100
# For example: land use
default_lu <- rep(swbm_lutype['Alfalfa','Code'], nfields)

lu_df <- swbm_build_field_value_df(nfields, start_date, end_date, default_lu)

# Make update table
updates <- data.frame('ID'=1:nfields, 'landuse_11'=swbm_lutype['Pasture','Code'])

# Do updates
lu_df <- swbm_landuse_update(lu_df, updates)

}
