% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ts_obs_sim_combine}
\alias{ts_obs_sim_combine}
\title{Time Series Observation & Simulation Dataframe CombineR}
\usage{
ts_obs_sim_combine(
  obs_df_list,
  sim_df_list,
  group_names,
  date_cols = c("Date", "Date"),
  val_cols,
  col_rename = T,
  save_cols = c()
)
}
\arguments{
\item{obs_df_list}{list of observation data frames}

\item{sim_df_list}{list of simulation data frames}

\item{group_names}{char array of names corresponding to obs-sim dataframe pairs. Intended
to allow datasets to be subset easily after the merge}

\item{date_cols}{char array of date column names for obs and sim dataframes, respectively.
Default: c('Date','Date')}

\item{val_cols}{char array of value column names for obs and sim dataframes, respectively}

\item{col_rename}{bool, rename value columns obs and sim (downside: user may lose track of units).
Default: TRUE}

\item{save_cols}{char array, names of extra columns to preserve in merge. Default: empty.}
}
\value{
dataframe of Date, Obs, Sim, group and save_cols
}
\description{
Time Series Observation & Simulation Dataframe CombineR
}
\note{
identical val columns between sim/obs datasets likely will fail during the subset step
since the second val_col will likely be renamed during the date merge
}
