% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMODFLOW.R
\name{update_DIS_stress_periods}
\alias{update_DIS_stress_periods}
\title{Write SVIHM MODFLOW Discretization File (DIS) with Updated Stress Periods}
\usage{
update_DIS_stress_periods(
  num_steps,
  num_stress_periods,
  output_dir,
  filename = "SVIHM.dis",
  ref_data_dir = data_dir["ref_data_dir", "loc"],
  verbose = TRUE
)
}
\arguments{
\item{num_steps}{Number of time steps (i.e. days) in each stress period}

\item{num_stress_periods}{Number of stress periods}

\item{output_dir}{Directory to write DIS file in}

\item{filename}{character name of file to write (default: SVIHM.dis)}

\item{ref_data_dir}{Reference data directory containing DIS template file, SVIHM_dis_reference.txt
(default: SVIHM/SVIHM_Input_Files/reference_data/)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None, writes file to disk
}
\description{
Uses the reference SVIHM DIS file as a template and write a new version with the updated
temporal discretization at the specified directory.
}
\examples{
num_days <- days_in_month_diff(scen$start_date, scen$end_date)
update_DIS_stress_periods(num_steps=num_days, num_stress_periods=length(num_days), '.')
}
\author{
Claire Kouba
}
