% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMODFLOW.R
\name{update_DRN_stress_periods}
\alias{update_DRN_stress_periods}
\title{Write SVIHM MODFLOW Drain File (DRN) with Updated Number of Stress Periods}
\usage{
update_DRN_stress_periods(
  num_stress_periods,
  output_dir,
  filename = "SVIHM.drn",
  default_drn_value = 10000,
  ref_data_dir = data_dir["ref_data_dir", "loc"],
  verbose = TRUE
)
}
\arguments{
\item{num_stress_periods}{Number of stress periods}

\item{output_dir}{Directory to write DRN file in}

\item{filename}{character name of file to write (default: SVIHM.drn)}

\item{default_drn_value}{Numeric value to be written for drain conductance}

\item{ref_data_dir}{Reference data directory containing DIS template file, SVIHM_dis_reference.txt
(default: SVIHM/SVIHM_Input_Files/reference_data/)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None, writes file to disk
}
\description{
In SVIHM, the drains represent saturation excess overland flow within the discharge zone. The
drain activation elevations are set at land surface.
}
\examples{
update_DRN_stress_periods(416, '.')

}
\author{
Claire Kouba
}
