% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMODFLOW.R
\name{update_DRNO_stress_periods}
\alias{update_DRNO_stress_periods}
\title{Write SVIHM MODFLOW Drain Overland Flow File (DRNO) with Updated Number of Stress Periods}
\usage{
update_DRNO_stress_periods(
  num_stress_periods,
  output_dir,
  filename = "SVIHM.drno",
  default_drn_value = NA,
  ref_data_dir = data_dir["ref_data_dir", "loc"],
  verbose = TRUE
)
}
\arguments{
\item{num_stress_periods}{Number of stress periods}

\item{output_dir}{Directory to write DRN file in}

\item{filename}{character name of file to write (default: SVIHM.drno)}

\item{default_drn_value}{Numeric value to be written for drain conductance (default, NA, results in 10,000)}

\item{ref_data_dir}{Reference data directory (default: SVIHM/SVIHM_Input_Files/reference_data/)
with \code{drno_cells.txt} file that defines drain cells and SFR segments}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None, writes file to disk
}
\description{
In SVIHM, the drains represent saturation excess overland flow within the discharge zone. The
drain activation elevations are set at land surface. The DRNO package is a modified version of
the MODFLOW drain package able to route water "removed" from the MODFLOW model through the drain
package to the SFR package. Each drain cell, therefore, has an associated SFR segment for it
to route water to.
}
\examples{
update_DRNO_stress_periods(416, '.')

}
\author{
Leland Scantlebury, Claire Kouba
}
