% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMODFLOW.R
\name{update_OC_stress_periods}
\alias{update_OC_stress_periods}
\title{Write SVIHM MODFLOW Output Control File (OC) with Updated Number of Stress Periods}
\usage{
update_OC_stress_periods(
  num_steps,
  num_stress_periods,
  output_dir,
  filename = "SVIHM.oc",
  monthly = T,
  save_budget = T,
  save_drawdown = T,
  save_heads = T,
  verbose = TRUE
)
}
\arguments{
\item{num_steps}{Number of time steps (i.e. days) in each stress period}

\item{num_stress_periods}{Number of stress periods}

\item{output_dir}{Directory to write OC file in}

\item{filename}{character name of file to write (default: SVIHM.hob)}

\item{monthly}{T/F Write print/save statements only at month ends (as opposed to at every time step, F)}

\item{save_budget}{T/F write out Cell-by-cell budget file (very large) (default: True)}

\item{save_drawdown}{T/F write out drawdown file (default: True)}

\item{save_heads}{T/F write out head file (default: True)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\description{
Write SVIHM MODFLOW Output Control File (OC) with Updated Number of Stress Periods
}
\author{
Claire Kouba, Leland Scantlebury
}
