% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_ag_pumping_file}
\alias{write_ag_pumping_file}
\title{Write file specifying pumping volumes in agricultural wells}
\usage{
write_ag_pumping_file(
  start_date,
  n_stress,
  output_dir,
  ag_pumping_data = NA,
  filename = "ag_well_specified_volume.txt"
)
}
\arguments{
\item{scenario_name}{Name of  management scenario. Default is historical basecase or "basecase".}
}
\value{
none; writes input file
}
\description{
Write file specifying pumping volumes in agricultural wells
}
\examples{
# Dates
start_date <- get_model_start(1991)
end_date <- as.Date(floor_date(Sys.Date(), 'month')-1)
# Fields
nfields <- 100
# For example: land use
default_irr <- rep(swbm_irrtype['Wheel Line','Code'], nfields)

lu_df <- swbm_build_field_value_df(nfields, start_date, end_date, default_irr)

# Make update table - magically all fields updated to center pivot in 2005!
updates <- data.frame('ID'=1:nfields, 'Year'=2005)

# Do updates
lu_df <- swbm_irrtype_cp_update(lu_df, updates)

}
