% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_daily_crop_coeff_values_file}
\alias{write_daily_crop_coeff_values_file}
\title{Write file specifying crop coefficients for each crop type, for each daily timestep}
\usage{
write_daily_crop_coeff_values_file(
  model_start_date,
  model_end_date,
  output_dir
)
}
\arguments{
\item{model_start_date}{Start date of simulation}

\item{model_end_date}{End date of simulation}

\item{output_dir}{directory to write the files in}
}
\value{
none; writes file
}
\description{
Write file specifying crop coefficients for each crop type, for each daily timestep
}
