% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_sfr_network_file}
\alias{write_sfr_network_file}
\title{Write SWBM SFR Network File}
\usage{
write_sfr_network_file(
  nsteps,
  output_dir,
  filename = "SFR_network.txt",
  daily = FALSE,
  ntabfiles = 30,
  comment = "SFR Package written by RSVP"
)
}
\arguments{
\item{nsteps}{Total number of MODFLOW time steps (i.e. days in model) to be found in tab files}

\item{output_dir}{directory to write the files to}

\item{filename}{output filename (default: sfr_network.txt)}

\item{daily}{T/F if using daily time steps in SFR (i.e., tabfiles) (default: FALSE)}

\item{ntabfiles}{if daily, the number of tabfiles to be read by MODFLOW (default: 30)}

\item{comment}{SFR file comment printed at top of file (default: 'SFR Package written by RSVP')}
}
\value{
none; writes file
}
\description{
Writes the file that essentially gets used as the SFR package template in the SWBM. Has support
for including daily tabfiles.
}
