% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{write_streamflow_by_subws_input_file}
\alias{write_streamflow_by_subws_input_file}
\title{Write Surface Water Input File (one combined record per subwatershed)}
\usage{
write_streamflow_by_subws_input_file(
  gauges = NA,
  tribs_df = NA,
  output_dir,
  start_date,
  end_date,
  monthly = T,
  filename = "streamflow_input.txt",
  verbose = TRUE
)
}
\arguments{
\item{gauges}{List of gauge data from \code{\link{get_tributary_flows}}}

\item{output_dir}{Directory to write file}

\item{start_date}{Simulation start date}

\item{end_date}{Simulation end date}

\item{filename}{Filename (optional, default: streamflow_input.txt)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None
}
\description{
Write Surface Water Input File (one combined record per subwatershed)
}
