% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression_old.R
\name{write_streamflow_input_file}
\alias{write_streamflow_input_file}
\title{Write Streamflow Input File}
\usage{
write_streamflow_input_file(
  stream_out,
  output_dir,
  filename = "streamflow_input.txt",
  verbose = TRUE
)
}
\arguments{
\item{stream_out}{DataFrame, output from generate_streamflow_input()}

\item{output_dir}{Directory to write file}

\item{filename}{Filename (optional, default: streamflow_input.txt)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None
}
\description{
Write Streamflow Input File
}
