% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMODFLOW.R
\name{write_SVIHM_head_obs_file}
\alias{write_SVIHM_head_obs_file}
\title{Write SVIHM Head Observation (HOBS) File}
\usage{
write_SVIHM_head_obs_file(
  model_start_date,
  model_end_date,
  output_dir,
  filename = "SVIHM.hob",
  ref_data_dir = data_dir["ref_data_dir", "loc"],
  verbose = TRUE
)
}
\arguments{
\item{model_start_date}{Date of model start}

\item{model_end_date}{Date of model end}

\item{output_dir}{Directory to write HOBS file in}

\item{filename}{character name of file to write (default: SVIHM.hob)}

\item{ref_data_dir}{Reference data directory containing DIS template file, SVIHM_dis_reference.txt
(default: SVIHM/SVIHM_Input_Files/reference_data/)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{

}
\description{
TODO: Accept data as arguments? Automate new data download?
}
\author{
Claire Kouba
}
