% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_crop_coefficient_file}
\alias{write_SWBM_crop_coefficient_file}
\title{Write SWBM Crop Coefficient (Kc) File}
\usage{
write_SWBM_crop_coefficient_file(kc_df, output_dir, filename, verbose = TRUE)
}
\arguments{
\item{kc_df}{Dataframe of days and coefficients (see
\code{\link{gen_daily_binary_crop_coefficients}} and
\code{\link{gen_daily_curve_crop_coefficients}})}

\item{output_dir}{directory to write the files to}

\item{filename}{general input filename, instream_flow_available_ratio.txt by default}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\description{
Write SWBM Crop Coefficient (Kc) File
}
