% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_drain_files}
\alias{write_SWBM_drain_files}
\title{Write SWBM Drains Input Files (Drain_m3day.txt, Drains_initial_m3day.txt)}
\usage{
write_SWBM_drain_files(
  num_stress_periods,
  output_dir,
  values = NULL,
  drains_filename = "Drains_m3day.txt",
  init_filename = "Drains_initial_m3day.txt",
  verbose = TRUE
)
}
\arguments{
\item{num_stress_periods}{Number of model stress periods, determining the length of values array
to be written to the drains file}

\item{output_dir}{directory to write the files in}

\item{values}{Values to write to drain files, by default (NULL) will write zeroes}

\item{drains_filename}{Optional, default "Drains_m3day.txt"}

\item{verbose}{T/F write status info to console (default: TRUE)}

\item{drains_iniital_filename}{Optional, default "Drains_initial_m3day.txt"}
}
\value{
None
}
\description{
Write SWBM Drains Input Files (Drain_m3day.txt, Drains_initial_m3day.txt)
}
