% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_gen_inputs_file}
\alias{write_SWBM_gen_inputs_file}
\title{Write SWBM General Inputs File}
\usage{
write_SWBM_gen_inputs_file(
  output_dir,
  num_stress_periods,
  filename = "general_inputs.txt",
  modelName = "SVIHM",
  WYstart = 1991,
  npoly = 2119,
  nlandcover = 6,
  nAgWells = 167,
  nMuniWells = 0,
  nSubws = 8,
  inflow_is_vol = FALSE,
  daily_sw = TRUE,
  nSFR_inflow_segs = 12,
  nrows = 440,
  ncols = 210,
  RD_Mult = 1.4,
  calib_software = "UCODE",
  using_neighbor_irr_rule = TRUE,
  scenario_id = "basecase",
  irr_ditch_file = "irr_ditch.txt",
  verbose = TRUE
)
}
\arguments{
\item{output_dir}{directory to write the files to}

\item{num_stress_periods}{Number of model stress periods}

\item{filename}{general input filename, general_inputs.txt by default}

\item{verbose}{T/F write status info to console (default: TRUE)}

\item{recharge_scenario}{character}

\item{flow_scenario}{character}

\item{alf_irr_stop_mo}{integer Calendar month}

\item{alf_irr_stop_day}{integer}

\item{early_cutoff_flag}{character}

\item{curtailment_scenario}{character}

\item{curtail_start_mo}{integer}

\item{curtail_start_day}{integer}

\item{landuse_scenario}{character}
}
\value{
None
}
\description{
Writes all basecase values by default
}
