% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_gen_inputs_file_orig}
\alias{write_SWBM_gen_inputs_file_orig}
\title{Write SWBM General Inputs File (DEPRECIATED)}
\usage{
write_SWBM_gen_inputs_file_orig(
  num_stress_periods,
  output_dir,
  filename = "general_inputs.txt",
  recharge_scenario = "Basecase",
  flow_scenario = "Basecase",
  alf_irr_stop_mo = 8,
  alf_irr_stop_day = 31,
  early_cutoff_flag = "AllYears",
  curtailment_scenario = "NoCurtail",
  curtail_start_mo = 8,
  curtail_start_day = 15,
  landuse_scenario = "basecase",
  verbose = TRUE
)
}
\arguments{
\item{num_stress_periods}{Number of model stress periods}

\item{output_dir}{directory to write the files to}

\item{filename}{general input filename, general_inputs.txt by default}

\item{recharge_scenario}{character}

\item{flow_scenario}{character}

\item{alf_irr_stop_mo}{integer Calendar month}

\item{alf_irr_stop_day}{integer}

\item{early_cutoff_flag}{character}

\item{curtailment_scenario}{character}

\item{curtail_start_mo}{integer}

\item{curtail_start_day}{integer}

\item{landuse_scenario}{character}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None
}
\description{
Writes all basecase values by default
}
