% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_instream_available_file}
\alias{write_SWBM_instream_available_file}
\title{Write SWBM Instream Flow Available Ratio File}
\usage{
write_SWBM_instream_available_file(
  avail_monthly,
  output_dir,
  filename = "instream_flow_available_ratio.txt",
  verbose = TRUE
)
}
\arguments{
\item{avail_monthly}{Dataframe of ratio of instream flow available over the model period, as
generated by \code{\link{build_cdfw_instream_flow_cal}}}

\item{output_dir}{directory to write the files to}

\item{filename}{input filename, instream_flow_available_ratio.txt by default}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\description{
Write SWBM Instream Flow Available Ratio File
}
