% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{write_swbm_precip_factor_file}
\alias{write_swbm_precip_factor_file}
\title{Write SWBM Precipitation Factor File}
\usage{
write_swbm_precip_factor_file(
  output_dir,
  filename = "precip_factors.txt",
  precip_factors_df,
  verbose = TRUE
)
}
\arguments{
\item{output_dir}{Directory to write file}

\item{filename}{Precip factors filename}

\item{precip_factors_df}{A two-column dataframe containing the SWBM polygon (field) ID; and the
precip multiplication factor.}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None
}
\description{
Writes the Soil Water Balance Model (SWBM) precipitation factor file (precip_factors.txt).
Consists of one factor per land use polygon (i.e., field). Is used to multiply the
daily precip value to implement spatial precipitation gradients.
}
\author{
Leland Scantlebury and Claire Kouba
}
