% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipData.R
\name{write_swbm_precip_input_file}
\alias{write_swbm_precip_input_file}
\title{Write SWBM Precipitation Input File}
\usage{
write_swbm_precip_input_file(
  p_record,
  output_dir,
  merge_dataset = NULL,
  filename = "precip_regressed.txt",
  verbose = TRUE
)
}
\arguments{
\item{p_record}{DataFrame of precipitation data, as generated by get_daily_precip_table()
Expects, at a minimum, columns "stitched" and "Date"}

\item{output_dir}{Directory to write file}

\item{merge_dataset}{DataFrame with columns "Date" and "PRCP" to use to overwrite values in p_record.
All non-NA values will be used to overwrite p_record}

\item{filename}{Filename (optional, default: precip_regressed.txt)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None
}
\description{
Writes the Soil Water Balance Model (SWBM) precipitation input file (precip.txt). Allows for a
another time series data frame to serve as an "overwrite" for the data via a merge. This is to
try to accomodate different regression analyses used throughout time.
}
\details{
Uses the "stitched" column of the p_record as the final output (after merge_dataset merge)
}
\author{
Leland Scantlebury
}
