% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_SFR_diversions_file}
\alias{write_SWBM_SFR_diversions_file}
\title{Write SFR Diversions File}
\usage{
write_SWBM_SFR_diversions_file(
  filename = "SFR_diversions.txt",
  output_dir,
  num_divs = 2,
  iseg_for_divs = c(3, 10),
  iprior_for_divs = c(1, 1),
  flow_for_divs = c(19600, 39100),
  verbose = T
)
}
\arguments{
\item{num_divs}{Number of diversions in the stream network}

\item{iseg_for_divs}{The stream segment associated with each diversion}

\item{iprior_for_divs}{Specifies priority handling for diversion in MODFLOW.}

\item{flow_for_divs}{Uniform flowrate for the diversion.}
}
\description{
Generates text file specifying the number, stream segment, priority handling (in MODFLOW), and
unchanging flowrate of each diversion point in the stream network.
Default is currently (Nov 2022) configured to specify diversion points for the SVID ditch and
Farmers ditch, with 0 flow in them (as ditch diversions are currently  handled in a
separate module).
}
\author{
Claire Kouba, Leland Scantlebury
}
