% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_SFR_files}
\alias{write_SWBM_SFR_files}
\title{Write SWBM File Partitioning Subwatershed Surface Flows to Stream Inflows}
\usage{
write_SWBM_SFR_files(sfr_component, output_dir, filename, verbose = TRUE)
}
\arguments{
\item{sfr_component}{Dataframe of months and subwatershed inflow partition fractions (see
\code{\link{gen_monthly_sfr_flow_partition}})}

\item{output_dir}{directory to write the files to}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\description{
Write SWBM File Partitioning Subwatershed Surface Flows to Stream Inflows
}
