% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_SFR_inflow_files}
\alias{write_SWBM_SFR_inflow_files}
\title{Write SWBM File Partitioning Subwatershed Surface Flows to Stream Inflows}
\usage{
write_SWBM_SFR_inflow_files(
  sfr_component,
  output_dir,
  filename,
  verbose = TRUE
)
}
\arguments{
\item{sfr_component}{Dataframe of months and either: 1) subwatershed inflow partition
fractions (see \code{\link{gen_monthly_sfr_flow_partition}}), or 2) inflows available
for irrigation, or 3) inflows NOT available for irrigation (e.g. reserved for environmental
flows) (for items 2 and 3 see \code{\link{process_monthly_sfr_inflows}}).}

\item{output_dir}{directory to write the files to}

\item{filename}{Writes to this filename}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\description{
Write SWBM File Partitioning Subwatershed Surface Flows to Stream Inflows
}
