% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{write_tributary_input_file_2_forks_old}
\alias{write_tributary_input_file_2_forks_old}
\title{Write Tributary Input File - Two Forks Version (old)}
\usage{
write_tributary_input_file_2_forks_old(
  gauges,
  output_dir,
  start_date,
  end_date,
  filename = "streamflow_input.txt",
  verbose = TRUE
)
}
\arguments{
\item{gauges}{List of gauge data from \code{\link{get_tributary_flows}}}

\item{output_dir}{Directory to write file}

\item{start_date}{Simulation start date}

\item{end_date}{Simulation end date}

\item{filename}{Filename (optional, default: streamflow_input.txt)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
None
}
\description{
Write Tributary Input File - Two Forks Version (old)
}
