% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{add_FJ_to_dflist}
\alias{add_FJ_to_dflist}
\title{Add FJ (standardized) Streamflow to dataframe to use as regression dependant variable}
\usage{
add_FJ_to_dflist(gaugelist, FJ_id = 1, ...)
}
\arguments{
\item{FJ_id}{}
}
\value{

}
\description{
Add FJ (standardized) Streamflow to dataframe to use as regression dependant variable
}
