% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{alter_SWBM_sfr_inflows}
\alias{alter_SWBM_sfr_inflows}
\title{Alter subwatershed inflow time series according to scenario name}
\usage{
alter_SWBM_sfr_inflows(
  subws_inflows,
  scenario_id,
  min_flow_file_name = NA,
  rerun_FJ_trib_corr = F
)
}
\arguments{
\item{subws_inflows}{Wide data table of inflow data for each day (rows) for each tributary (column)}

\item{scenario_id}{Management scenario identifier}
}
\value{
None, writes file to disk
}
\description{
Alter subwatershed inflow time series according to scenario name
}
\examples{
create_SWBM_sfr_inflows_df(subws_inflows, scen$name)
}
\author{
Claire Kouba, Leland Scantlebury
}
