% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETData.R
\name{build_daily_et_df}
\alias{build_daily_et_df}
\title{Build Daily ET DataFrame}
\usage{
build_daily_et_df(
  start_date,
  end_date,
  ref_data_dir = data_dir["ref_data_dir", "loc"],
  verbose = TRUE
)
}
\arguments{
\item{start_date}{Start of data period}

\item{end_date}{End of data period}

\item{ref_data_dir}{Reference data directory
(optional, default: SVIHM/SVIHM_Input_Files/reference_data)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\value{
dataframe of ET data from start_date to end_date
}
\description{
Combines 3 datasets to create a final evapotranspiration dataset for SVIHM:
- Original SVIHM ET data (ref_et_1991_2011.txt)
- CIMIS Station 225 (Scott Valley) measured ET (from 2015-04-19 to current)
- Spatial SVIHM ET data for gap between datasets
}
\author{
Claire Kouba, Leland Scantlebury
}
