% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{complete_ts}
\alias{complete_ts}
\title{Complete monthly time series by filling with NA values}
\usage{
complete_ts(
  df,
  date_col = "Date",
  by = "month",
  start_date = NULL,
  end_date = NULL,
  drop_extra = T
)
}
\arguments{
\item{drop_extra}{}
}
\description{
Complete monthly time series by filling with NA values
}
