% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETData.R
\name{download_cimis_data}
\alias{download_cimis_data}
\title{Download CIMIS Station Data}
\usage{
download_cimis_data(start_date, end_date, stations = 225, verbose = TRUE, ...)
}
\arguments{
\item{start_date}{data request start date}

\item{end_date}{data request end date}

\item{stations}{CIMIS stations (default is 225, Scott Valley)
#' @param verbose T/F write status info to console (default: TRUE)}

\item{...}{additional arguments passed to \code{\link[cimir]{cimis_data}}}
}
\value{

}
\description{

}
\details{
CIMIS key must be set by \code{\link[cimir]{set_key}} before using this function.

More options are available from the full \code{\link[cimir]{cimis_data}} function.
}
\author{
Leland Scantlebury
}
