% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{get_fj_flow_file_name}
\alias{get_fj_flow_file_name}
\title{Get Fort Jones Flow File Name}
\usage{
get_fj_flow_file_name(start_date, end_date)
}
\arguments{
\item{end_date}{}
}
\value{

}
\description{
Codifies a pattern for naming
}
