% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.precip.compare}
\alias{plot.precip.compare}
\title{Precipitation Comparison Bar Chart}
\usage{
\method{plot}{precip.compare}(
  dates,
  precip1,
  precip2,
  unit = "mm",
  title = NULL,
  colors = NULL,
  legend_names
)
}
\arguments{
\item{dates}{array of dates of precip events}

\item{precip1}{array of first precipitation dataset at dates}

\item{precip2}{array of second precipitation dataset at dates}

\item{unit}{character, the units the precipitation is measured in}

\item{title}{title of plot (optional)}

\item{colors}{colors for data (identical, precip1, precip2) (optional)}

\item{legend_names}{names added to legend to describe precip1 and precip2}
}
\description{
Creates plot showing identical (color 1) and differing (color 2, color 3) precipitation data
}
