% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.precip.cumulative}
\alias{plot.precip.cumulative}
\title{Title}
\usage{
\method{plot}{precip.cumulative}(dates, ..., title = NULL, unit = "mm", col = NULL)
}
\arguments{
\item{col}{}
}
\value{

}
\description{
Title
}
