% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_swbm_volumetric_budget}
\alias{plot_swbm_volumetric_budget}
\title{Plot Volumetric Budget for SWBM}
\usage{
plot_swbm_volumetric_budget(
  swbm_dir,
  out_dir,
  plot_type = c("monthly", "yearly"),
  write_csv = TRUE,
  swbm_terms = c("Precipitation", "SW Irrigation", "GW Irrigation", "ET", "Recharge",
    "Runoff", "Storage", "Error"),
  swbm_colors = c("lightblue1", "blueviolet", "midnightblue", "goldenrod", "chartreuse4",
    "mistyrose", "orangered3", "gray"),
  origin_date = as.Date("1990-09-30")
)
}
\arguments{
\item{swbm_dir}{Character string. Directory path where the SWBM files are located.}

\item{out_dir}{Character string. Directory path where the output plots should be saved.}

\item{plot_type}{Character string. Type of plot to generate, either "monthly" or "yearly". Defaults to "monthly".}

\item{write_csv}{Logical. If TRUE, writes the SWBM data to a CSV file. Defaults to TRUE.}

\item{swbm_terms}{Character vector. The terms (components) of the SWBM volumetric budget to plot. Defaults to c('Precipitation', 'SW Irrigation', 'GW Irrigation', 'ET', 'Recharge', 'Runoff', 'Storage', 'Error').}

\item{swbm_colors}{Character vector. Colors for each term in the plot. Defaults to c('lightblue1', 'darkcyan', 'midnightblue', 'goldenrod', 'green4', 'mistyrose', 'black', 'gray').}

\item{origin_date}{Date. The starting date for the SWBM model run. Defaults to as.Date('1990-09-30').}
}
\value{
No return value. The function generates and saves the plot files.
}
\description{
This function generates plots for the Soil Water Balance Model (SWBM) volumetric budget, either
on a monthly (line plot) or yearly (bar plot) basis.
}
\details{
The function reads the SWBM output file \code{monthly_water_budget.dat}, processes the data, and
generates a bar plot for the annual volumetric budget or a line plot for the monthly budget.
The user can switch between monthly and yearly plots using the \code{plot_type} argument.
}
\examples{
\dontrun{
# Generate yearly volumetric budget plot:
plot_swbm_volumetric_budget(swbm_dir = 'path_to_swbm_dir', out_dir = 'path_to_output_dir', plot_type = "yearly")

# Generate monthly volumetric budget plot:
plot_swbm_volumetric_budget(swbm_dir = 'path_to_swbm_dir', out_dir = 'path_to_output_dir', plot_type = "monthly")
}

}
