% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverRegression.R
\name{predict_and_destandardize}
\alias{predict_and_destandardize}
\title{Title}
\usage{
predict_and_destandardize(gauge, lm_object, lm_nodata = NULL, use_obs = TRUE)
}
\arguments{
\item{use_obs}{(T/F) overwrite predictions with recorded (observed) data}
}
\description{
Title
}
