% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSVP-package.R
\docType{package}
\name{RSVP-package}
\alias{RSVP}
\alias{RSVP-package}
\title{RSVP: R Scott Valley Package (RSVP)}
\description{
R functions for pre-processing, post-processing, and scenario development for the Scott Valley Integrated Hydrologic Model (SVIHM)
}
\author{
\strong{Maintainer}: Leland Scantlebury \email{lscantle@ucdavis.edu}

Authors:
\itemize{
  \item Claire Kouba \email{cmkouba@ucdavis.edu}
}

}
\keyword{internal}
