% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_curtailment_file}
\alias{write_SWBM_curtailment_file}
\title{Write file specifying irrigation curtailment regulations/practices}
\usage{
write_SWBM_curtailment_file(scenario_name = "basecase", start_date, n_stress)
}
\arguments{
\item{scenario_name}{Name of  management scenario. Default is historical basecase or "basecase".}
}
\value{
curtail_tab
}
\description{
Write file specifying irrigation curtailment regulations/practices
}
