% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_curtailment_file}
\alias{write_SWBM_curtailment_file}
\title{Write file specifying Managed Aquifer Recharge (applied irrigation) volumes for each field, for each stress period}
\usage{
write_SWBM_curtailment_file(
  scenario_id = "basecase",
  output_dir,
  start_date,
  end_date,
  apn_overlap_threshold = 0.5,
  exclude_terrible_and_major_overflow_polygons = F,
  curtail_100_pct_non_lcs_fields = T
)
}
\arguments{
\item{scenario_name}{Name of  management scenario. Default is historical basecase or "basecase".}
}
\value{
none, saves MAR_volumes.txt
}
\description{
Write file specifying Managed Aquifer Recharge (applied irrigation) volumes for each field, for each stress period
}
