% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_curtailment_file}
\alias{write_SWBM_curtailment_file}
\title{Write file specifying curtailment for each field, for each stress period}
\usage{
write_SWBM_curtailment_file(output_dir, start_date, end_date)
}
\arguments{
\item{output_dir}{directory to write the files in}

\item{start_date}{Start date of simulation}

\item{end_date}{End date of simulation}
}
\value{
none, saves curtailment_fractions.txt
}
\description{
Write file specifying curtailment for each field, for each stress period
}
