% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETData.R
\name{write_swbm_et_input_file}
\alias{write_swbm_et_input_file}
\title{Write SWBM Evapotranspiration Input File}
\usage{
write_swbm_et_input_file(
  et_record,
  output_dir,
  filename = "ref_et_monthly.txt",
  verbose = TRUE
)
}
\arguments{
\item{et_record}{DataFrame of ET data, as generated by \code{\link{build_daily_et_df}}
Expects, at a minimum, columns "ETo_mm" and "Date"}

\item{output_dir}{Directory to write file}

\item{filename}{Filename (optional, default: ref_et_monthly.txt)}

\item{verbose}{T/F write status info to console (default: TRUE)}
}
\description{
Writes the Soil Water Balance Model (SWBM) evapotranspiration (ET) input file (ref_et_monthly.txt).
}
\examples{
\dontrun{
model_start_date <- get_model_start(1991)
model_end_date <- as.Date(floor_date(Sys.Date(), 'month')-1)
update_dir <- create_update_dir(end_date = model_end_date)

et <- build_daily_et_df(model_start_date, model_end_date)
write_swbm_et_input_file(et_record = et,
                         output_dir = update_dir,
                         filename = 'ref_et.txt')
}
}
\seealso{
\code{\link{build_daily_et_df}}
}
\author{
Claire Kouba, Leland Scantlebury
}
