% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSWBMInputs.R
\name{write_SWBM_main_input_file}
\alias{write_SWBM_main_input_file}
\title{Write SWBM Main Input File (svihm.swbm)}
\usage{
write_SWBM_main_input_file(
  output_dir,
  num_stress_periods,
  filename = "svihm.swbm",
  modelName = "SVIHM",
  WYstart = 1991,
  npoly = 2119,
  nlandcover = 6,
  nAgWells = 167,
  nSpecWells = 0,
  nSubws = 8,
  inflow_is_vol = FALSE,
  daily_sw = TRUE,
  nSFR_inflow_segs = 12,
  nrows = 440,
  ncols = 210,
  neighborRuleYearDay = 250,
  absoluteIrrDate = c(5, 15),
  writeUCODE = TRUE,
  writePEST = FALSE,
  precip_file = "precip.txt",
  poly_landcover_file = "polygon_landcover_ids.txt",
  poly_agwell_file = "ag_well_list_by_polygon.txt",
  et_file = "ref_et.txt",
  et_ext_depth_file = "ET_Cells_Extinction_Depth.txt",
  ets_template_file = "SVIHM_ETS_template.txt",
  kc_frac_file = "kc_values.txt",
  sfr_network_file = "SFR_network.txt",
  sfr_partition_file = "SFR_subws_flow_partitioning.txt",
  wel_template_file = "SVIHM_WEL_template.txt",
  agwell_locs_file = "ag_well_summary.txt",
  recharge_zones_file = "recharge_zones.txt",
  sfr_jtf_file = "SFR_network_jtf.txt",
  irr_ditch_file = "irr_ditch.txt",
  et_zone_cells_file = "ET_Zone_Cells.txt",
  mar_depth_file = "MAR_depth.txt",
  curtail_frac_file = "curtailment_fractions.txt",
  et_correction_file = "field_et_corrections.txt",
  water_mover_file = NULL,
  specwell_locs_file = NULL,
  specwell_vol_file = NULL,
  print_daily_fields = list(),
  verbose = TRUE
)
}
\arguments{
\item{output_dir}{Directory where the input file will be written.}

\item{num_stress_periods}{Integer. Number of model stress periods (NMONTHS in the input file).}

\item{filename}{Character. The name of the output file. Default is "svihm.swbm".}

\item{modelName}{Character. The name of the MODFLOW model. Default is "SVIHM".}

\item{WYstart}{Integer. The water year start (WYSTART in the input file). Default is 1991.}

\item{npoly}{Integer. The number of polygons (NPOLY in the input file). Default is 2119.}

\item{nlandcover}{Integer. The number of land cover types (NLANDCOVER in the input file). Default is 6.}

\item{nAgWells}{Integer. The number of agricultural wells (NAGWELLS in the input file). Default is 167.}

\item{nSubws}{Integer. The number of sub-watersheds (NSUBWS in the input file). Default is 8.}

\item{inflow_is_vol}{Logical. Whether the inflow is volume-based. Default is FALSE.}

\item{daily_sw}{Logical. Whether to write daily soil water budget (DAILY_SW in the input file). Default is TRUE.}

\item{nSFR_inflow_segs}{Integer. The number of streamflow routing inflow segments (NSFR_INFLOW_SEGS in the input file). Default is 12.}

\item{nrows}{Integer. The number of rows in the model grid (NROWS in the input file). Default is 440.}

\item{ncols}{Integer. The number of columns in the model grid (NCOLS in the input file). Default is 210.}

\item{neighborRuleYearDay}{Integer. The neighbor rule cutoff day of the year (NEIGHBOR_RULE in the input file). Default is 250.}

\item{absoluteIrrDate}{Integer vector of length 2. The absolute irrigation cutoff date as month and day (ABSOLUTE_IRR_DATE in the input file). Default is c(5, 15).}

\item{writeUCODE}{Logical. Whether to write UCODE files (WRITE_UCODE in the input file). Default is TRUE.}

\item{writePEST}{Logical. Whether to write PEST files (WRITE_PEST in the input file). Default is FALSE.}

\item{precip_file}{Character. The name of the precipitation input file (PRECIP in the input file). Default is "precip.txt".}

\item{poly_landcover_file}{Character. The name of the polygon landcover IDs file (POLY_LANDCOVER in the input file). Default is "polygon_landcover_ids.txt".}

\item{et_file}{Character. The name of the reference evapotranspiration file (ET in the input file). Default is "ref_et.txt".}

\item{et_ext_depth_file}{Character. The name of the ET cells extinction depth file (ET_EXT_DEPTH in the input file). Default is "ET_Cells_Extinction_Depth.txt".}

\item{ets_template_file}{Character. The name of the ETS template file (ETS_TEMPLATE in the input file). Default is "SVIHM_ETS_template.txt".}

\item{kc_frac_file}{Character. The name of the crop coefficient fractions file (KC_FRAC in the input file). Default is "kc_values.txt".}

\item{sfr_network_file}{Character. The name of the streamflow routing network file (SFR_NETWORK in the input file). Default is "SFR_network.txt".}

\item{sfr_partition_file}{Character. The name of the streamflow routing partition file (SFR_PARTITION in the input file). Default is "SFR_subws_flow_partitioning.txt".}

\item{wel_template_file}{Character. The name of the well template file (WEL_TEMPLATE in the input file). Default is "SVIHM_WEL_template.txt".}

\item{recharge_zones_file}{Character. The name of the recharge zones file (RECHARGE_ZONES in the input file). Default is "recharge_zones.txt".}

\item{irr_ditch_file}{Character. The name of the irrigation ditch file (IRR_DITCH in the input file). Default is "irr_ditch.txt".}

\item{et_zone_cells_file}{Character. The name of the ET zone cells file (ET_ZONE_CELLS in the input file). Default is "ET_Zone_Cells.txt".}

\item{mar_depth_file}{Character. The name of the managed aquifer recharge depth file (MAR_DEPTH in the input file). Default is "MAR_depth.txt".}

\item{curtail_frac_file}{Character. The name of the curtailment fractions file (CURTAIL_FRAC in the input file). Default is "curtailment_fractions.txt".}

\item{water_mover_file}{Character. The name of the water mover file (WATER_MOVER in the input file). Default is NULL.}

\item{print_daily_fields}{List of named lists. Each element should contain \code{id} (Integer, Field ID) and \code{prefix} (Character, filename prefix) for fields to print daily in the PRINT_DAILY block.}

\item{verbose}{Logical. Whether to print status information to the console. Default is TRUE.}

\item{nMuniWells}{Integer. The number of municipal wells (NMUNIWELLS in the input file). Default is 0.}

\item{RD_Mult}{Numeric. The root depth multiplier (RD_MULT in the input file). Default is 1.4.}
}
\value{
None. The function writes the input file to the specified directory.
}
\description{
This function writes the main input file for the Soil Water Budget Model (SWBM).
}
\examples{
# Set Directory (Created in SVIHM_Input_Files/Updates) - Grabs latest version
update_dir <- latest_dir(data_dir['update_dir','loc'])

# Set output fields
print_daily_fields <- list(
  list(id = 1, prefix = "Field_1"),
  list(id = 91, prefix = "Field_91"),
  list(id = 136, prefix = "AG_MIX_Flood")
# Add more fields as needed

# Write file
write_SWBM_main_input_file(output_dir = update_dir, num_stress_periods = 395)
}
